+++ docs/changes.rst
@@ -6,6 +6,8 @@ Version 0.8.1
 
 To be released.
 
+- Added :mod:`sqlalchemy_imageattach.migration` module for storage migration.
+
 
 Version 0.8.0
 -------------